requirements.txt.txt
llama-index==0.8.59
langchain==0.0.329
PyPDF2==3.0.1
python-docx==1.0.1
streamlit=1.28.1
python-dotenv==1.0.0